avoid modifying CHUNK_SIZE
whisper was falling into an endless loop when running on two files in a
row due to the `CHUNK_SIZE` constant being modified at the end of the
first file. this adds a separate `chunk_dur` variable that can safely be
modified and is reset every time `start()` is called.